move invite event responder to core bot logic
With that, all modules should be now using the provided event consumer
loops, which should make moving sending messages to a central point
easier as well.

Slowly moving towards making modules *not* depend on running with a
logged-in matrix client, which Should™ make things at least easier to
write tests for.